+++ lib/features/home/home_minimal_screen.dart
@@ -390,4 +390,3 @@ class _ErrorView extends StatelessWidget {
     );
   }
 }
-